requirements.txt
numpy==1.23.5
torch==2.1.0
hydra-core==1.3.2
timm==1.0.12
tqdm==4.67.1
opencv-python
matplotlib
pandas
kagglehub
datasets
tensorboard